Add utils file and copy mechanism with fallback
